we no longer need httpclient.DumpInvalidPayload

+++ devices/fan.go
@@ -5,8 +5,8 @@ type Fan struct {
 	Status     string
 	Position   int
 	Model      string
-	CurrentRPM int
-	MaxRPM     int
-	MinRPM     int
+	CurrentRPM int64
+	MaxRPM     int64
+	MinRPM     int64
 	PowerKw    float64
 }
